+++ src/theme/material-ui.ts
@@ -1,0 +1,48 @@
+import { createTheme, ThemeProvider as MuiThemeProvider } from '@mui/material/styles';
+
+// MUI v7 主题配置
+export const muiTheme = createTheme({
+  palette: {
+    primary: {
+      main: '#3f51b5',
+      light: '#7986cb',
+      dark: '#303f9f',
+      contrastText: '#ffffff',
+    },
+    secondary: {
+      main: '#f50057',
+      light: '#ff5983',
+      dark: '#c51162',
+      contrastText: '#ffffff',
+    },
+    error: {
+      main: '#f44336',
+      light: '#e57373',
+      dark: '#d32f2f',
+    },
+    warning: {
+      main: '#ff9800',
+      light: '#ffb74d',
+      dark: '#f57c00',
+    },
+    info: {
+      main: '#2196f3',
+      light: '#64b5f6',
+      dark: '#1976d2',
+    },
+    success: {
+      main: '#4caf50',
+      light: '#81c784',
+      dark: '#388e3c',
+    },
+    background: {
+      default: '#fafafa',
+      paper: '#ffffff',
+    },
+    text: {
+      primary: 'rgba(0, 0, 0, 0.87)',
+      secondary: 'rgba(0, 0, 0, 0.6)',
+      disabled: 'rgba(0, 0, 0, 0.38)',
+    },
+  },
+});
