--- src/utils/theme.ts
@@ -1,261 +1,0 @@
-// src/config/theme.ts
-import { createTheme, ThemeProvider as MuiThemeProvider } from '@mui/material/styles';
-import { CssBaseline, GlobalStyles } from '@mui/material';
-
-// MUI v7 主题配置
-const baseTheme = createTheme({
-  colorSchemes: {
-    light: {
-      palette: {
-        primary: {
-          main: '#3f51b5',
-          light: '#7986cb',
-          dark: '#303f9f',
-          contrastText: '#ffffff',
-        },
-        secondary: {
-          main: '#f50057',
-          light: '#ff5983',
-          dark: '#c51162',
-          contrastText: '#ffffff',
-        },
-        error: {
-          main: '#f44336',
-          light: '#e57373',
-          dark: '#d32f2f',
-        },
-        warning: {
-          main: '#ff9800',
-          light: '#ffb74d',
-          dark: '#f57c00',
-        },
-        info: {
-          main: '#2196f3',
-          light: '#64b5f6',
-          dark: '#1976d2',
-        },
-        success: {
-          main: '#4caf50',
-          light: '#81c784',
-          dark: '#388e3c',
-        },
-        background: {
-          default: '#fafafa',
-          paper: '#ffffff',
-        },
-        text: {
-          primary: 'rgba(0, 0, 0, 0.87)',
-          secondary: 'rgba(0, 0, 0, 0.6)',
-          disabled: 'rgba(0, 0, 0, 0.38)',
-        },
-      },
-    },
-    dark: {
-      palette: {
-        primary: {
-          main: '#90caf9',
-          light: '#e3f2fd',
-          dark: '#42a5f5',
-          contrastText: '#000000',
-        },
-        secondary: {
-          main: '#f48fb1',
-          light: '#fce4ec',
-          dark: '#e91e63',
-          contrastText: '#000000',
-        },
-        error: {
-          main: '#f44336',
-          light: '#e57373',
-          dark: '#d32f2f',
-        },
-        warning: {
-          main: '#ff9800',
-          light: '#ffb74d',
-          dark: '#f57c00',
-        },
-        info: {
-          main: '#2196f3',
-          light: '#64b5f6',
-          dark: '#1976d2',
-        },
-        success: {
-          main: '#4caf50',
-          light: '#81c784',
-          dark: '#388e3c',
-        },
-        background: {
-          default: '#121212',
-          paper: '#1e1e1e',
-        },
-        text: {
-          primary: '#ffffff',
-          secondary: 'rgba(255, 255, 255, 0.7)',
-          disabled: 'rgba(255, 255, 255, 0.5)',
-        },
-      },
-    },
-  },
-  typography: {
-    fontFamily: [
-      'Inter',
-      'Roboto',
-      '-apple-system',
-      'BlinkMacSystemFont',
-      '"Segoe UI"',
-      '"Helvetica Neue"',
-      'Arial',
-      'sans-serif',
-    ].join(','),
-    h1: {
-      fontSize: '2.5rem',
-      fontWeight: 700,
-      lineHeight: 1.2,
-    },
-    h2: {
-      fontSize: '2rem',
-      fontWeight: 600,
-      lineHeight: 1.3,
-    },
-    h3: {
-      fontSize: '1.75rem',
-      fontWeight: 600,
-      lineHeight: 1.3,
-    },
-    h4: {
-      fontSize: '1.5rem',
-      fontWeight: 600,
-      lineHeight: 1.4,
-    },
-    h5: {
-      fontSize: '1.25rem',
-      fontWeight: 600,
-      lineHeight: 1.4,
-    },
-    h6: {
-      fontSize: '1.125rem',
-      fontWeight: 600,
-      lineHeight: 1.4,
-    },
-    body1: {
-      fontSize: '1rem',
-      lineHeight: 1.6,
-    },
-    body2: {
-      fontSize: '0.875rem',
-      lineHeight: 1.5,
-    },
-    button: {
-      textTransform: 'none',
-      fontWeight: 500,
-    },
-  },
-  shape: {
-    borderRadius: 8,
-  },
-  components: {
-    MuiButton: {
-      styleOverrides: {
-        root: ({ theme }) => ({
-          borderRadius: 8,
-          textTransform: 'none',
-          fontWeight: 500,
-          fontSize: '0.875rem',
-          minHeight: 36,
-          transition: 'all 0.2s ease-in-out',
-          '&:hover': {
-            transform: 'translateY(-1px)',
-            boxShadow: theme.shadows[4],
-          },
-        }),
-        containedPrimary: ({ theme }) => ({
-          background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
-          '&:hover': {
-            background: 'linear-gradient(135deg, #5a6fd8 0%, #6a4190 100%)',
-          },
-        }),
-      },
-    },
-    MuiCard: {
-      styleOverrides: {
-        root: ({ theme }) => ({
-          borderRadius: 12,
-          transition: 'all 0.2s ease-in-out',
-          '&:hover': {
-            transform: 'translateY(-2px)',
-            boxShadow: theme.shadows[8],
-          },
-        }),
-      },
-    },
-    MuiPaper: {
-      styleOverrides: {
-        root: {
-          borderRadius: 12,
-        },
-      },
-    },
-    MuiTextField: {
-      styleOverrides: {
-        root: {
-          '& .MuiOutlinedInput-root': {
-            borderRadius: 8,
-            transition: 'all 0.2s ease-in-out',
-            '&:hover': {
-              '& .MuiOutlinedInput-notchedOutline': {
-                borderWidth: 2,
-              },
-            },
-            '&.Mui-focused': {
-              '& .MuiOutlinedInput-notchedOutline': {
-                borderWidth: 2,
-              },
-            },
-          },
-        },
-      },
-    },
-    MuiListItemButton: {
-      styleOverrides: {
-        root: {
-          borderRadius: 8,
-          margin: '2px 4px',
-          transition: 'all 0.2s ease-in-out',
-          '&:hover': {
-            transform: 'translateX(4px)',
-          },
-          '&.Mui-selected': {
-            transform: 'translateX(4px)',
-            fontWeight: 600,
-          },
-        },
-      },
-    },
-    MuiChip: {
-      styleOverrides: {
-        root: {
-          borderRadius: 16,
-          fontWeight: 500,
-        },
-      },
-    },
-    MuiAlert: {
-      styleOverrides: {
-        root: {
-          borderRadius: 8,
-        },
-      },
-    },
-    MuiAppBar: {
-      styleOverrides: {
-        root: {
-          backdropFilter: 'blur(10px)',
-        },
-      },
-    },
-  },
-});
-
-export const theme = baseTheme;
-
-export default theme;
